+++ package.json
@@ -19,6 +19,7 @@
     "react-dom": "18.2.0",
     "react-native": "0.73.6",
     "react-native-calendars": "^1.1304.1",
+    "react-native-datepicker": "^1.7.2",
     "react-native-safe-area-context": "^4.10.3",
     "react-native-screens": "^3.31.1",
     "react-native-web": "~0.19.6"
